requirements.txt
aiogram>=3.3.0
aiohttp>=3.9.0
python-dotenv>=1.0.0
flask>=3.0.0
gunicorn>=21.0.0
